% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myMazeShinyApp.R
\name{myMazeShinyApp}
\alias{myMazeShinyApp}
\title{Building mazes app}
\usage{
myMazeShinyApp()
}
\value{
The function returns a shiny app with the possibility to build a maze
and see its solution.
}
\description{
Building mazes app
}
\details{
According to shiny settings, the app offers the user to build a maze by selecting the inputs
weight, height, and the Depth First Search algorithm.
}
\examples{
\dontrun{myMazeShinyApp()}
}
\author{
Joost and Edward
}
