% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myMazeShinyApp3.R
\name{myMazeShinyApp3}
\alias{myMazeShinyApp3}
\title{Building mazes app}
\value{
The function returns a shiny app with the possibility to build a maze
and see its solution.
}
\description{
Building mazes app
}
\details{
According to shiny settings, the app offers the user to build a maze by selecting the inputs
weight, height. Two algorithms are used : DFS and Recursive Backtracking
}
\examples{
\dontrun{myMazeShinyApp3()}
}
\author{
Joost and Edward
}
