% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmyMaze.R
\name{plotmyMaze}
\alias{plotmyMaze}
\title{ploymyMaze function}
\arguments{
\item{GraphMade}{An existing graph}

\item{nrows}{number of rows, maze height value arbitrarly set to 0.}

\item{ncols}{number of cols, maze weight value arbitrarly set to 0.}
}
\value{
The function returns a plot of the existing maze in 2D.
}
\description{
ploymyMaze function
}
\details{
The function transforms the graph into a data frame and then 
creates a plot of the existing maze with ggplot2 package.
}
\examples{
Example_maze <- RetakeExamF::myGraph (nrows = 5,ncols = 9) 
Example_maze <- RetakeExamF::dfs_method (GraphMade = Example_maze)
RetakeExamF::plotmyMaze(Example_maze,nrows= 5,ncols = 5)
}
\author{
Joost and Edward
}
