% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_maze.R
\name{rb_maze}
\alias{rb_maze}
\title{Joost_maze function}
\arguments{
\item{height}{number of rows, maze height value arbitrarly set to 0.}

\item{width}{number of cols, maze weight value arbitrarly set to 0.}
}
\value{
The function returns a plot of the existing maze in 2D.
}
\description{
Joost_maze function
}
\details{
The function uses the Recursive Backtracking algorithm to build a maze.
This algorithm starts by creating a grid of cells and then repeatedly selects a cell 
to carve a path while backtracking to create dead-ends.
}
\examples{
RetakeExamF:: rb_maze(31,31) 

}
\author{
Joost and Edward
}
